Update post review and enable smart camera connect in cli
